refactor: break state manager constant import cycle
