+++ commands/admin.py
@@ -15,7 +15,7 @@ from .building import (
 )
 from world.stats import CORE_STAT_KEYS, ALL_STATS
 from world.system import stat_manager
-from world.system.state_manager import MAX_LEVEL
+from world.system.constants import MAX_LEVEL
 from utils.stats_utils import get_display_scroll, normalize_stat_key
 from utils import VALID_SLOTS
