+++ world/system/__init__.py
@@ -2,5 +2,6 @@
 
 from . import state_manager
 from . import stat_manager
+from . import constants
 
-__all__ = ["state_manager", "stat_manager"]
+__all__ = ["state_manager", "stat_manager", "constants"]
