+++ world/system/constants.py
@@ -1,0 +1,9 @@
+# Constants used across the world.system package
+
+# Maximum fullness a character can reach from food or drink.
+MAX_SATED = 100
+
+# Highest level a player character can reach.
+MAX_LEVEL = 100
+
+__all__ = ["MAX_SATED", "MAX_LEVEL"]
